+++ Lib/fontbakery/profiles/hhea.py
@@ -8,35 +8,6 @@ profile_imports = [
     ('.shared_conditions', ('glyph_metrics_stats', 'is_ttf'))
 ]
 
-def _get_missing_tables(req_tables_set, ttFont):
-    """Returns a sorted list of table tags not supported by the font."""
-    return sorted(req_tables_set - set(ttFont.keys()))
-
-
-@check(
-    id = 'com.google.fonts/check/linegaps',
-    proposal = 'legacy:check/041'
-)
-def com_google_fonts_check_linegaps(ttFont):
-    """Checking Vertical Metric Linegaps."""
-    required_tables = {"hhea", "OS/2"}
-    missing_tables = _get_missing_tables(required_tables, ttFont)
-    if missing_tables:
-        for table_tag in missing_tables:
-            yield FAIL, Message("lacks-table", f"Font lacks '{table_tag}' table.")
-        return
-
-    if ttFont["hhea"].lineGap != 0:
-        yield WARN,\
-              Message("hhea",
-                      "hhea lineGap is not equal to 0.")
-    elif ttFont["OS/2"].sTypoLineGap != 0:
-        yield WARN,\
-              Message("OS/2",
-                      "OS/2 sTypoLineGap is not equal to 0.")
-    else:
-        yield PASS, "OS/2 sTypoLineGap and hhea lineGap are both 0."
-
 
 @check(
     id = 'com.google.fonts/check/maxadvancewidth',
@@ -44,8 +15,9 @@ def com_google_fonts_check_linegaps(ttFont):
 )
 def com_google_fonts_check_maxadvancewidth(ttFont):
     """MaxAdvanceWidth is consistent with values in the Hmtx and Hhea tables?"""
+
     required_tables = {"hhea", "hmtx"}
-    missing_tables = _get_missing_tables(required_tables, ttFont)
+    missing_tables = sorted(required_tables - set(ttFont.keys()))
     if missing_tables:
         for table_tag in missing_tables:
             yield FAIL,\