+++ Lib/fontbakery/profiles/universal.py
@@ -75,6 +75,7 @@ UNIVERSAL_PROFILE_CHECKS = \
         'com.google.fonts/check/whitespace_widths',
         'com.google.fonts/check/interpolation_issues',
         'com.google.fonts/check/math_signs_width',
+        'com.google.fonts/check/linegaps',
     ]
 
 
@@ -2163,6 +2164,47 @@ def com_google_fonts_check_math_signs_width(ttFont):
                       f"{outliers_summary}")
     else:
         yield PASS, "Looks good."
+
+
+@check(
+    id = 'com.google.fonts/check/linegaps',
+    rationale = """
+        The LineGap value is a space added to the line height created by the union
+        of the (typo/hhea)Ascender and (typo/hhea)Descender. It is handled differently
+        according to the environment.
+
+        This leading value will be added above the text line in most desktop apps.
+        It will be shared above and under in web browsers, and ignored in Windows
+        if Use_Typo_Metrics is disabled.
+
+        For better linespacing consistency across platforms,
+        (typo/hhea)LineGap values must be 0.
+    """,
+    proposal = ['https://github.com/googlefonts/fontbakery/issues/4133',
+                'https://googlefonts.github.io/gf-guide/metrics.html']
+)
+def com_google_fonts_check_linegaps(ttFont):
+    """Checking Vertical Metric Linegaps."""
+
+    required_tables = {"hhea", "OS/2"}
+    missing_tables = sorted(required_tables - set(ttFont.keys()))
+    if missing_tables:
+        for table_tag in missing_tables:
+            yield FAIL,\
+                  Message("lacks-table",
+                          f"Font lacks '{table_tag}' table.")
+        return
+
+    if ttFont["hhea"].lineGap != 0:
+        yield FAIL,\
+              Message("hhea",
+                      "hhea lineGap is not equal to 0.")
+    elif ttFont["OS/2"].sTypoLineGap != 0:
+        yield FAIL,\
+              Message("OS/2",
+                      "OS/2 sTypoLineGap is not equal to 0.")
+    else:
+        yield PASS, "OS/2 sTypoLineGap and hhea lineGap are both 0."
 
 
 profile.auto_register(globals())